+++ aws/sqs/src/main/scala/com/nike/fleam/sqs/SqsDelete.scala
@@ -1,6 +1,8 @@
 package com.nike.fleam
 package sqs
 
+import cats.Show
+import cats.implicits._
 import com.nike.fleam.configuration._
 import sqs.configuration._
 import com.amazonaws.services.sqs.model._
@@ -8,6 +10,7 @@ import com.amazonaws.services.sqs.AmazonSQSAsync
 import akka.stream.scaladsl._
 import ContainsMessage.ops._
 import com.nike.fawcett.sqs._
+import Keyed.ops._
 
 import scala.concurrent.{ExecutionContext, Future}
 
@@ -57,13 +60,27 @@ class SqsDelete(
       deleteMessage(request)
     }
 
-    def batched[T: ContainsMessage](ts: List[T])(implicit ec: ExecutionContext): Future[BatchResult[T]] = {
+    /** Deletes batches of messages
+     *
+     *  @tparam T Input type that contains a message and provides a key to track deletes.
+     *            Due to the at-least-once nature of SQS it's possible to have batches with duplicate messages ids.
+     *            Providing an alternate unique id in your type can avoid this problem if desired.
+     *            Keyed and ContainsMessage for Message can be imported from com.nike.fleam.sqs.implicits._
+     *  @tparam Key Type of Key used for tracking deletes. It must provide an instance of Show.
+     *              For Messages use type `MessageId` with the default implicits provided to use the Amazon provided
+     *              message id.
+     *              Import cats.implicits._ to get a `Show` instance for `MessageId`.
+     *  @param ts A list of `T` items to batch delete. Must be 10 or less items. For automatic grouping use `toFlow` instead.
+     *  @param ec ExecutionContext
+     *  @return BatchResult containing AWS DeleteMessageBatchResult and results divided into successful and failed.
+     */
+    def batched[T: ContainsMessage : Keyed[?, Key], Key : Show](ts: List[T])(implicit ec: ExecutionContext): Future[BatchResult[T]] = {
       if (ts.isEmpty) {
         Future.successful(BatchResult(new DeleteMessageBatchResult(), Nil, Nil))
       } else {
         val entries = ts.map { t =>
           val message = t.getMessage
-          new DeleteMessageBatchRequestEntry().withId(message.getMessageId).withReceiptHandle(message.getReceiptHandle)
+          new DeleteMessageBatchRequestEntry().withId(t.getKey.show).withReceiptHandle(message.getReceiptHandle)
         }
 
         val request = modifyBatchRequest { new DeleteMessageBatchRequest()
@@ -76,22 +93,36 @@ class SqsDelete(
             BatchResult(
               deleteMessageBatchResult = deleteMessageBatchResult,
               failed = DeleteMessageBatchResultLens.failed.get(deleteMessageBatchResult).map(result =>
-                ts.find(_.getMessage.getMessageId == result.getId).map(FailedResult(_, result))
+                ts.find(_.getKey.show == result.getId).map(FailedResult(_, result))
               ).flatten,
               successful = DeleteMessageBatchResultLens.successful.get(deleteMessageBatchResult).map(result =>
-                ts.find(_.getMessage.getMessageId == result.getId).map(SuccessfulResult(_, result))
+                ts.find(_.getKey.show == result.getId).map(SuccessfulResult(_, result))
               ).flatten
             )
           }
       }
     }
 
-    def toFlow[T: ContainsMessage](config: SqsProcessingConfiguration)(implicit ec: ExecutionContext):
+    /** Deletes batches of messages
+     *
+     *  @tparam T Input type that contains a message and provides a key to track deletes.
+     *            Due to the at-least-once nature of SQS it's possible to have batches with duplicate messages ids.
+     *            Providing an alternate unique id in your type can avoid this problem if desired.
+     *            Keyed and ContainsMessage for Message can be imported from com.nike.fleam.sqs.implicits._
+     *  @tparam Key Type of Key used for tracking deletes. It must provide an instance of Show.
+     *              For Messages use type `MessageId` with the default implicits provided to use the Amazon provided
+     *              message id.
+     *              Import cats.implicits._ to get a `Show` instance for `MessageId`.
+     *  @param config options for parallelism and batchsize
+     *  @param ec ExecutionContext
+     *  @return Flow from T to BatchResult containing AWS DeleteMessageBatchResult and results divided into successful and failed.
+     */
+    def toFlow[T: ContainsMessage : Keyed[?, Key], Key: Show](config: SqsProcessingConfiguration)(implicit ec: ExecutionContext):
         Flow[T, BatchResult[T], akka.NotUsed] =
       Flow[T]
         .via(config.groupedWithin.toFlow)
         .map(_.toList)
-        .mapAsync(config.parallelism) { batched[T] }
+        .mapAsync(config.parallelism) { batched[T, Key] }
   }
 
   def forQueue(url: String): UrlFixedQueue = new UrlFixedQueue(url)