+++ aws/sqs/src/main/scala/com/nike/fleam/sqs/SqsRetry.scala
@@ -87,7 +87,7 @@ object SqsRetry {
         SqsEnqueue(client, modifyBatchRequest = SqsEnqueue.delayMessagesBySeconds(delaySeconds))
           .forQueue(config.queue.url)
           .batched(messages)(implicitly[ToMessage[Message]], ec),
-      deleteMessages = SqsDelete(client).forQueue(config.queue.url).batched,
+      deleteMessages = SqsDelete(client).forQueue(config.queue.url).batched[Message, MessageId],
       deadLetterEnqueueMessages = (messages, ec) =>
         SqsEnqueue(client)
           .forQueue(config.deadLetterQueue.url)